+++ internal/command/terminal_echo_linux.go
@@ -1,0 +1,18 @@
+//go:build linux
+
+package command
+
+import "golang.org/x/sys/unix"
+
+func disableTerminalEcho(fd int) (func() error, error) {
+	state, err := unix.IoctlGetTermios(fd, unix.TCGETS)
+	if err != nil {
+		return nil, err
+	}
+	noEcho := *state
+	noEcho.Lflag &^= unix.ECHO
+	if err = unix.IoctlSetTermios(fd, unix.TCSETS, &noEcho); err != nil {
+		return nil, err
+	}
+	return func() error { return unix.IoctlSetTermios(fd, unix.TCSETS, state) }, nil
+}